fixing requested changes
