--- db/migrate/20170822144440_add_first_name_to_user.rb
@@ -1,5 +1,0 @@
-class AddFirstNameToUser < ActiveRecord::Migration[5.1]
-  def change
-    add_column :users, :first_name, :string
-  end
-end
