--- db/migrate/20170822144623_add_last_name_to_user.rb
@@ -1,5 +1,0 @@
-class AddLastNameToUser < ActiveRecord::Migration[5.1]
-  def change
-    add_column :users, :last_name, :string
-  end
-end
