--- db/migrate/20170822144702_add_date_of_birth_to_user.rb
@@ -1,5 +1,0 @@
-class AddDateOfBirthToUser < ActiveRecord::Migration[5.1]
-  def change
-    add_column :users, :date_of_birth, :datetime
-  end
-end
