--- db/migrate/20170822160301_add_hometown_to_user.rb
@@ -1,7 +1,0 @@
-class AddHometownToUser < ActiveRecord::Migration[5.1]
-  def change
-    add_column :users, :hometown, :string
-    add_column :users, :occupation, :string
-    add_column :users, :bio, :string
-  end
-end
