+++ spec/features/user_signup_info_spec.rb
@@ -1,4 +1,5 @@
 require 'rails_helper'
+require 'user_helpers.rb'
 
 feature "user" do
   let(:email) {'user@gmail.com'}
@@ -10,22 +11,8 @@ feature "user" do
   let(:occupation) {'Coach'}
   let(:bio) {"I'm pretty cool and stuff. Now go feed the MonsterCat"}
 
-
-
   scenario "I can input further information about myself on signing up" do
-    visit('/users/sign_up')
-    fill_in 'user_first_name', with: first_name
-    fill_in 'user_last_name', with: last_name
-    select 1890, :from => "user_date_of_birth_1i"
-    select "May", :from => "user_date_of_birth_2i"
-    select 1, :from => "user_date_of_birth_3i"
-    fill_in 'user_email', with: email
-    fill_in 'user_password', with: password
-    fill_in 'user_password_confirmation', with: password
-    fill_in 'user_hometown', with: hometown
-    fill_in 'user_occupation', with: occupation
-    fill_in 'user_bio', with: bio
-    click_button 'Sign up'
+    sign_up
     expect(page).to have_content('Welcome! You have signed up successfully.')
   end
 end
